(maint) adding the dependency on to pe-bolt-server
By adding the dependency on pe-bolt-server we can piggyback of the dependencies used as majority of them are identical.

This means we will mirror the versions released with bolt-server, but also means we no longer need to package the bolt dependencies.

Keeping puma as our own component so that we can still laucnh it from the ace-server bin vs the bolt-server bin.